Improve follow-on RSVP logic
When a guest submits an RSVP, a PUT request is sent to the server.
Prior to this commit, React handled the logic to update the guest-
list table that the user sees, but that was fraught  because the
client can not know the other guest's emails (or guest ID's).
Here, I remove the front-end logic and simply send all the updated
guest-list in a JSON response to the PUT request (RSVP). This
commit also sends the Identified guest's name to the page for
guests that have already RSVP'd so they don't need to re-type it.