Remove server.listen(), db.connect() from app.js
Supertest sends mock requests directly to the express app object, and
not the http server that is created by calling listen(), so we don't
need to start it for the tests.  Also, we want the tests files to determine
which db to use for tests.  By moving both these calls to a separate
file, we can access the express app for testing without enabling the
unnecessary http server and connecting to the wrong database.